move to rubygems-style project structure
