+++ Gemfile
@@ -1,8 +1,3 @@
-# frozen_string_literal: true
-source "https://rubygems.org"
+source 'https://rubygems.org'
 
-gem "rpairtree", require: "pairtree"
-
-group :test do
-  gem "rspec"
-end
+gemspec
