+++ lib/datasets.rb
@@ -1,0 +1,13 @@
+require "datasets/version"
+
+require "datasets/filter"
+require "datasets/filesystem"
+require "datasets/pairtree_path_resolver"
+require "datasets/path_resolver"
+require "datasets/scheduler"
+require "datasets/volume"
+require "datasets/volume_creator"
+require "datasets/volume_linker"
+require "datasets/volume_writer"
+require "datasets/zip_writer"
+
