+++ lib/datasets/filesystem.rb
