+++ lib/datasets/path_resolver.rb
