+++ lib/datasets/version.rb
@@ -1,0 +1,3 @@
+module Datasets
+  VERSION = "0.1.0"
+end
