+++ lib/datasets/volume_writer.rb
