+++ lib/datasets/zip_writer.rb
